+++ pytorch_lightning/trainer/data_loading.py
@@ -30,9 +30,11 @@ from pytorch_lightning.trainer.states import RunningStage
 from pytorch_lightning.trainer.supporters import CombinedLoader
 from pytorch_lightning.utilities import _TORCH_GREATER_EQUAL_1_6, rank_zero_warn
 from pytorch_lightning.utilities.apply_func import apply_to_collection
+from pytorch_lightning.utilities.auto_restart import _sampler_metadata_collate
 from pytorch_lightning.utilities.data import has_iterable_dataset, has_len
 from pytorch_lightning.utilities.debugging import InternalDebugger
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
+from pytorch_lightning.utilities.imports import _fault_tolerant_enabled
 from pytorch_lightning.utilities.model_helpers import is_overridden
 from pytorch_lightning.utilities.seed import pl_worker_init_function
 
@@ -259,6 +261,10 @@ class TrainerDataLoadingMixin(ABC):
         # add worker_init_fn for correct seeding in worker processes
         apply_to_collection(self.train_dataloader, DataLoader, self.auto_add_worker_init_fn)
 
+        # add collate_fn to collect metadata for fault tolerant training
+        if _fault_tolerant_enabled():
+            apply_to_collection(self.train_dataloader, DataLoader, self._add_sampler_metadata_collate)
+
         # wrap the sequence of train loaders to a CombinedLoader object for computing the num_training_batches
         self.train_dataloader = CombinedLoader(self.train_dataloader, self.data_connector.multiple_trainloader_mode)
 
@@ -460,9 +466,6 @@ class TrainerDataLoadingMixin(ABC):
     def request_dataloader(self, model: 'pl.LightningModule', stage: str) -> DataLoader:
         """Handles downloading data in the GPU or TPU case.
 
-        Args:
-            dataloader_fx: The bound dataloader getter
-
         Returns:
             The dataloader
         """
@@ -474,11 +477,16 @@ class TrainerDataLoadingMixin(ABC):
 
     def _flatten_dl_only(self, dataloaders):
         # handles user error when they return:
-        # return dl1, dl2  vs  return (dl1, dl2)
-        if isinstance(dataloaders, tuple):
-            all_dls = [isinstance(x, Iterable) for x in dataloaders]
-            all_dls = all(all_dls)
-            if all_dls:
-                dataloaders = list(dataloaders)
-
+        # `return dl1, dl2` vs `return (dl1, dl2)`
+        if isinstance(dataloaders, tuple) and all(isinstance(x, Iterable) for x in dataloaders):
+            return list(dataloaders)
         return dataloaders
+
+    @staticmethod
+    def _add_sampler_metadata_collate(dataloader: DataLoader) -> None:
+        """
+        Wrap default collate function to retrive ``FastForwardSampler`` state dict when fault tolerant is enabled.
+        """
+        dataloader.collate_fn = partial(
+            _sampler_metadata_collate, dataset=dataloader.dataset, default_collate=dataloader.collate_fn
+        )